Replace oc login with kube config
Using oc login causes issues with what we would like to do in the future
with respect to user impersonation, etc. Simply having a
`/opt/apb/.kube/config file` makes it so that our APBs will run in the
cluster as we expect them to (since the service account token and CA are
in known locations). This also allows for us to run APBs as sytem:admin
using docker as we continue to investigate APBs w/o an Ansible Broker.